reorganize and document a little
