+++ src/ocean_emulators/viz/__main__.py
@@ -1,0 +1,3 @@
+from ocean_emulators.viz.config import VizConfig, main
+
+main(VizConfig.from_yaml_and_cli())
